fix: Renaming to WebDidDocRouter
